Second commit to rename src/util/github.ts -> src/util/Github.ts
